Run an R expression and monitor memory

Evaluate an expression, and monitor the memory every poll_interval. This can monitor either the memory of your R expression and any processes it spawns or it can monitor the memory of your entire system.

with_monitor(
  expr,
  mode = "process",
  poll_interval = 0.1,
  monitor_file = tempfile(),
  overwrite = FALSE,
  gc_first = TRUE
)

Arguments

expr

The expression to evaluate.

mode

The monitor mode to use, process to monitor the function process and any spawned processes. system to monitor the memory of the entire system.

poll_interval

Time between each check of memory use in seconds.

monitor_file

Path to log memory monitor to, you can use this if your expr errors to recover the output for later plotting. Uses a tempfile by default.

overwrite

If TRUE then overwrite monitor_file if it exists already.

gc_first

If TRUE run a garbage collect before starting monitoring.

Value

A memprof_monitor object containing the result as the outcome from the expression and memory_use the memory used

Details

Note if using process monitor on windows it is best to set a long poll_interval of ~ 5s. This is because getting all the child processes of a process is quite a slow and CPU intensive task on windows meaning it will slow down execution if run too frequently.